Better VMRC functionality
VMRC and a number of other virtual consoles don't work with unicode sendinput, only with scancodes.

Switched to use scancodes by default and falling back to unicode input when we can't deduce the scancode.

Also we will now try to click in the middle of the VMRC window so the console should get the keyboard input.